+++ test/testdata/plain.sql
@@ -1,0 +1,45 @@
+--
+-- PostgreSQL database dump
+--
+-- Dumped from database version 12.4
+-- Dumped by pg_dump version 12.4
+SET statement_timeout = 0;
+
+SET lock_timeout = 0;
+
+SET idle_in_transaction_session_timeout = 0;
+
+SET client_encoding = 'UTF8';
+
+SET standard_conforming_strings = ON;
+
+SELECT
+    pg_catalog.set_config('search_path', '', FALSE);
+
+SET check_function_bodies = FALSE;
+
+SET xmloption = content;
+
+SET client_min_messages = warning;
+
+SET row_security = OFF;
+
+DROP EXTENSION IF EXISTS "uuid-ossp";
+
+--
+-- Name: uuid-ossp; Type: EXTENSION; Schema: -; Owner: -
+--
+CREATE EXTENSION IF NOT EXISTS "uuid-ossp" WITH SCHEMA public;
+
+--
+-- Name: EXTENSION "uuid-ossp"; Type: COMMENT; Schema: -; Owner:
+--
+COMMENT ON EXTENSION "uuid-ossp" IS 'generate universally unique identifiers (UUIDs)';
+
+SET default_tablespace = '';
+
+SET default_table_access_method = heap;
+
+--
+-- PostgreSQL database dump complete
+--